# Conflicts:
#	android/project/app/src/main/java/com/chill/mallang/ui/feature/fort_detail/FortDetailScreen.kt
#	android/project/app/src/main/java/com/chill/mallang/ui/feature/fort_detail/layout/GameStartBody.kt
#	android/project/app/src/main/java/com/chill/mallang/ui/navigation/MallangNavHost.kt

+++ android/project/app/src/main/java/com/chill/mallang/data/model/request/GradingUserAnswerRequest.kt
@@ -6,9 +6,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GradingUserAnswerRequest(
     @SerialName("quizId")
-    val quizId: Int,
+    val quizId: Long,
     @SerialName("userId")
-    val userId: Int,
+    val userId: Long,
     @SerialName("userAnswer")
     val userAnswer: String,
     @SerialName("answerTime")
